Splitting the API requests in another class
