+++ base/src/main/java/io/plaidapp/base/designernews/DesignerNewsPrefs.java
@@ -62,7 +62,6 @@ public class DesignerNewsPrefs {
     }
 
     public DesignerNewsService getApi() {
-        return loginRepository.getApi();
+        return loginRepository.getService();
     }
-
 }
